+++ src/react-components/src/components/02-molecules/Cards/EditionCard.stories.tsx
@@ -2,6 +2,8 @@ import * as React from "react";
 import bem from "../../../utils/bem";
 
 import EditionCard from "./EditionCard";
+import  IconLink from "../../01-atoms/Links/IconLink";
+import BasicLink from "../../01-atoms/Links/BasicLink";
 
 export default {
   title: "Edition Card",
@@ -13,9 +15,9 @@ export const editionCardWithFullData = () => <EditionCard
   blockName=""
   coverUrl="https://placeimg.com/300/400/arch"
   editionHeadingElement={<a href="edition-link" className={bem("link", [], "heading")} >2004 Edition</a>}
-  editionInfo={["Published in New York by Random House", "Written in English", "Under Creative Commons License"]}
-  readOnlineLink="#readOnlineUrl"
-  downloadLink="#downloadUrl">
+  editionInfo={[<>Published in New York by Random House</>, <>Written in English</>, <>Under Creative Commons License</>]}
+  readOnlineLink={<BasicLink className={bem("card-button-link", [], "edition-card")} url="blah">Read Online</BasicLink>}
+  downloadLink={<IconLink iconName="download" iconPosition="left" iconModifiers={["icon-left"]} url="blah">Download</IconLink>}>
 </EditionCard>;
 
 export const editionCardWithOneLinks = () => <EditionCard
@@ -23,8 +25,8 @@ export const editionCardWithOneLinks = () => <EditionCard
   blockName=""
   coverUrl="https://placeimg.com/300/400/arch"
   editionHeadingElement={<a href="edition-link" className={bem("link", [], "heading")} >2004 Edition</a>}
-  editionInfo={["Published in New York by Random House", "Written in English", "Under Creative Commons License"]}
-  readOnlineLink="#readOnlineUrl">
+  editionInfo={[<>Published in New York by Random House</>, <>Written in English</>, <>Under Creative Commons License</>]}
+  downloadLink={<IconLink iconName="download" iconPosition="left" iconModifiers={["icon-left"]} url="blah">Download</IconLink>}>
 </EditionCard>;
 
 export const editionCardMissingLinks = () => <EditionCard
@@ -32,7 +34,7 @@ export const editionCardMissingLinks = () => <EditionCard
   blockName=""
   coverUrl="https://placeimg.com/300/400/arch"
   editionHeadingElement={<a href="edition-link" className={bem("link", [], "heading")} >2004 Edition</a>}
-  editionInfo={["Published in New York by Random House", "Written in English", "Under Creative Commons License"]}
+  editionInfo={[<>Published in New York by Random House</>, <>Written in English</>, <>Under Creative Commons License</>]}
 >
 </EditionCard>;
 
@@ -41,19 +43,9 @@ export const editionCardWithSomeEditionInfo = () => <EditionCard
   blockName=""
   coverUrl="https://placeimg.com/300/400/arch"
   editionHeadingElement={<a href="edition-link" className={bem("link", [], "heading")} >2004 Edition</a>}
-  editionInfo={["Published in New York", "Under Creative Commons License"]}
-  readOnlineLink="#readOnlineUrl"
-  downloadLink="#downloadUrl">
-</EditionCard>;
-
-export const editionCardWithComponentInfo = () => <EditionCard
-  id="card#1"
-  blockName=""
-  coverUrl="https://placeimg.com/300/400/arch"
-  editionHeadingElement={<a href="edition-link" className={bem("link", [], "heading")} >2004 Edition</a>}
-  editionInfo={[<span>Published in New York</span>, "Under Creative Commons License"]}
-  readOnlineLink={<span>Read Online</span>}
-  downloadLink="#downloadUrl">
+  editionInfo={[<>Published in New York</>, <span>Under Creative Commons License</span>]}
+  readOnlineLink={<BasicLink className={bem("card-button-link", [], "edition-card")} url="blah">Read Online</BasicLink>}
+  downloadLink={<IconLink iconName="download" iconPosition="left" iconModifiers={["icon-left"]} url="blah">Download</IconLink>}>
 </EditionCard>;
 
 export const editionCardWithEmptyEditionInfo = () => <EditionCard
@@ -62,8 +54,8 @@ export const editionCardWithEmptyEditionInfo = () => <EditionCard
   coverUrl="https://placeimg.com/300/400/arch"
   editionHeadingElement={<a href="edition-link" className={bem("link", [], "heading")} >2004 Edition</a>}
   editionInfo={[]}
-  readOnlineLink="#readOnlineUrl"
-  downloadLink="#downloadUrl">
+  readOnlineLink={<BasicLink className={bem("card-button-link", [], "edition-card")} url="blah">Read Online</BasicLink>}
+  downloadLink={<IconLink iconName="download" iconPosition="left" iconModifiers={["icon-left"]} url="blah">Download</IconLink>}>
 </EditionCard>;
 
 export const editionCardWithAtypicalImageSize = () => <EditionCard
@@ -71,7 +63,17 @@ export const editionCardWithAtypicalImageSize = () => <EditionCard
   blockName=""
   coverUrl="https://placeimg.com/450/100/arch"
   editionHeadingElement={<a href="edition-link" className={bem("link", [], "heading")} >2004 Edition</a>}
-  editionInfo={["Publish Place: New York", "Language: English", "License: Creative Commons"]}
-  readOnlineLink="#readOnlineUrl"
-  downloadLink="#downloadUrl">
+  editionInfo={[<>Publish Place: New York</>, <>Language: English</>, <>License: Creative Commons</>]}
+  readOnlineLink={<BasicLink className={bem("card-button-link", [], "edition-card")} url="blah">Read Online</BasicLink>}
+  downloadLink={<IconLink iconName="download" iconPosition="left" iconModifiers={["icon-left"]} url="blah">Download</IconLink>}>
+</EditionCard>;
+
+export const instanceCard = () => <EditionCard
+  id="card#1"
+  blockName=""
+  coverUrl="https://placeimg.com/450/100/arch"
+  editionHeadingElement={<a href="edition-link" className={bem("link", [], "heading")} >2004 Edition</a>}
+  editionInfo={[<>Published in New York by Random House"</>, <a href="find-link" className={bem("link", [], "heading")}>Find in a library</a>]}
+  readOnlineLink={<BasicLink className={bem("card-button-link", [], "edition-card")} url="blah">Read Online</BasicLink>}
+  downloadLink={<IconLink iconName="download" iconPosition="left" iconModifiers={["icon-left"]} url="blah">Download</IconLink>}>
 </EditionCard>;